LayoutTests/imported/w3c/web-platform-tests/cors/redirect-preflight-2-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?origin=*&ident=fail_1489438122639 due to access control checks.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?headers=custom-header&origin=*&token=6de9061c-4017-44fc-a211-22e46653293f due to access control checks.
Preflight after redirect


PASS Same-origin custom-header request, redirect to cross-origin fails after doing a non-successful preflight 
FAIL Same-origin custom-header request, redirect to cross-origin succeeds after doing a preflight assert_unreached: Error during request Reached unreachable code
